Jenkins added and configured using sub-generator of jhpister

+++ src/main/java/com/mycompany/store/service/ProductService.java
@@ -1,12 +1,19 @@
 package com.mycompany.store.service;
 
+import com.mycompany.store.domain.OrderItem;
 import com.mycompany.store.domain.Product;
+import com.mycompany.store.repository.OrderItemRepository;
 import com.mycompany.store.repository.ProductRepository;
+
+import java.util.List;
 import java.util.Optional;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,8 +28,11 @@ public class ProductService {
 
     private final ProductRepository productRepository;
 
-    public ProductService(ProductRepository productRepository) {
+    private final OrderItemRepository orderItemRepository;
+
+    public ProductService(ProductRepository productRepository, OrderItemRepository orderItemRepository) {
         this.productRepository = productRepository;
+        this.orderItemRepository = orderItemRepository;
     }
 
     /**
@@ -127,8 +137,51 @@ public class ProductService {
      *
      * @param id the id of the entity.
      */
+    // return object instead of void from delete or deleteObject custom method
     public void delete(Long id) {
         log.debug("Request to delete Product : {}", id);
-        productRepository.deleteById(id);
+        Optional<OrderItem> orderItem = orderItemRepository.findOneWithToOneRelationships(id);
+        //Optional<Product> product = productRepository.findOneWithToOneRelationships(id);
+        if (orderItem.isEmpty() != true) {
+            //later set product.isActive = false
+            productRepository.findById(id).map(existingProduct -> {
+                existingProduct.setIsActive(false);
+                return existingProduct;
+            }).map(productRepository::save);
+            log.debug("This product cant be deleted", id);
+        }
+        //query to check the count of orderItem related to this product
+        else {
+            productRepository.deleteById(id);
+        }
+
     }
+
+    public Optional<Product> disableProduct(Product product) {
+        log.debug("Request to disable the Product : {}", product);
+
+        return productRepository
+            .findById(product.getId())
+            .map(existingProduct -> {
+
+                if (product.getIsActive() != null) {
+                    existingProduct.setIsActive(product.getIsActive());
+                }
+
+                return existingProduct;
+            })
+            .map(productRepository::save);
+
+    }
+
+    /*List<Product> findAllWithEagerRelationships(Boolean isActive) {
+        return this.getAllActiveProductsList(isActive);
+    }
+
+    @Query("select product from Product product where product.isActive =:isActive")
+    Optional<Product> getAllActiveProductsList(@Param("id") Long id) {
+        return null;
+    }*/
+
+
 }
